requirement.txt
requests
beautifulsoup4
google-generativeai